Schemas updated

+++ catapult-sdk/src/plugins/mosaic.js
@@ -32,14 +32,14 @@ const mosaicPlugin = {
 			id: ModelType.uint64HexIdentifier,
 			duration: ModelType.uint64,
 			nonce: ModelType.uint32,
-			flags: ModelType.int,
-			divisibility: ModelType.int
+			flags: ModelType.uint8,
+			divisibility: ModelType.uint8
 		});
 
 		builder.addTransactionSupport(EntityType.mosaicSupplyChange, {
 			mosaicId: ModelType.uint64HexIdentifier,
 			delta: ModelType.uint64,
-			action: ModelType.int
+			action: ModelType.uint8
 		});
 
 		builder.addSchema('mosaicDescriptor', {
@@ -53,8 +53,8 @@ const mosaicPlugin = {
 			startHeight: ModelType.uint64,
 			ownerAddress: ModelType.binary,
 			revision: ModelType.int,
-			flags: ModelType.int,
-			divisibility: ModelType.int,
+			flags: ModelType.uint8,
+			divisibility: ModelType.uint8,
 			duration: ModelType.uint64
 		});
 	},
